isProper graph function
Graph member function isProper to check if a set is proper

+++ src/visualisation.cpp
@@ -53,7 +53,7 @@ void drawParticle(SDL_Renderer* renderer, Vector center) {
 
 void compute_accelerations(Graph& graph, std::vector<Vector>& pos, std::vector<Vector>& acc, double l0) {
     // Get number of vertices
-    u64 N = graph.N;
+    int N = graph.N;
 
     // Compute the force between i and j
     for(int i=0 ; i<N ; i++) {
@@ -93,7 +93,7 @@ void compute_accelerations(Graph& graph, std::vector<Vector>& pos, std::vector<V
 
 void compute_positions(Graph& graph, std::vector<Vector>& pos, std::vector<Vector> vel, std::vector<Vector>& acc, int w, int h) {
     int limits = 20;
-    for (u64 i = 0; i<graph.N ; i++) {
+    for (int i = 0; i<graph.N ; i++) {
         double tmpX = pos[i].x;
         double tmpY = pos[i].y;
 
@@ -107,7 +107,7 @@ void compute_positions(Graph& graph, std::vector<Vector>& pos, std::vector<Vecto
 }
 
 void simulate(Graph& graph, std::vector<Vector>& pos, std::vector<Vector> vel, std::vector<Vector>& acc, double l0, int w, int h) {
-    u64 N = graph.N;
+    int N = graph.N;
     compute_accelerations(graph, pos, acc, l0);
     for(int i=0 ; i<N ; i++) {
         //vel[i].x = vel[i].x * 2;
@@ -124,15 +124,15 @@ void drawGraph(Graph& graph) {
 
     // Parameters
     int width = 800, height = 800;
-    u64 timeSteps = 300;
+    int timeSteps = 300;
     double l0 = 50;  // Spring size
-    u64 N = graph.N;
+    int N = graph.N;
 
     // Physicals variables
     std::vector<Vector> pos(N), vel(N), acc(N);
 
     // Initialize Random Positions and Velocities
-    for(u64 i=0 ; i<N ; i++) {
+    for(int i=0 ; i<N ; i++) {
         pos[i].x = randxy(200, 400);
         pos[i].y = randxy(200, 400);
 
@@ -147,15 +147,15 @@ void drawGraph(Graph& graph) {
     SDL_Init(SDL_INIT_VIDEO);
     SDL_CreateWindowAndRenderer(800, 800, SDL_WINDOW_OPENGL, &window, &renderer);
 
-    for(u64 t=0 ; t<timeSteps ; t++) {
+    for(int t=0 ; t<timeSteps ; t++) {
         simulate(graph, pos, vel, acc, l0, width, height);
 
         SDL_SetRenderDrawColor(renderer, 0, 0, 0, 255);
         SDL_RenderClear(renderer);
 
-        for(u64 i=0 ; i<N ; i++) {
+        for(int i=0 ; i<N ; i++) {
             drawParticle(renderer, pos[i]);
-            for(u64 j=0 ; j<N ; j++) 
+            for(int j=0 ; j<N ; j++) 
                 if (graph.areConnected(i,j))
                     SDL_RenderDrawLine(renderer, pos[i].x, pos[i].y, pos[j].x, pos[j].y);
         }
